+++ plugins/engine/renderloop/stdsteps/lightiter.h
@@ -142,8 +142,8 @@ public:
   virtual void Perform (iRenderView* rview, iSector* sector,
     csShaderVarStack &stacks);
 
-  virtual int AddStep (iRenderStep* step);
-  virtual int GetStepCount ();
+  virtual size_t AddStep (iRenderStep* step);
+  virtual size_t GetStepCount ();
 
   csPtr<iTextureHandle> GetAttenuationTexture (int attnType);
   csPtr<iTextureHandle> GetAttenuationTexture (const csVector3& attnVec);
