+++ plugins/sound/loader/au/aufile.cpp
@@ -57,7 +57,7 @@ public:
     SCF_DESTRUCT_IBASE();
   }
 
-  virtual csPtr<iSoundData> LoadSound(void *Buffer, uint32 Size);
+  virtual csPtr<iSoundData> LoadSound(void *Buffer, size_t Size);
 };
 
 SCF_IMPLEMENT_IBASE(csSoundLoader_AU)
@@ -82,7 +82,7 @@ SCF_IMPLEMENT_FACTORY(csSoundLoader_AU);
 #define Stream buf[index]
 
 csPtr<iSoundData>
-csSoundLoader_AU::LoadSound(void *databuf, uint32 size)
+csSoundLoader_AU::LoadSound(void *databuf, size_t size)
 {
   uint8 *buf = (uint8*) databuf;
   unsigned long index=0;
